Fix remaining float-based timeouts.

+++ src/model.ml
@@ -17,7 +17,7 @@ let need_admin authdata =
 type user_entry = {
   user_name: string;
   user_token: string option;
-  user_expires: float option;
+  user_expires: int64 option;
   user_alt_email: string option;
   user_level: level
 }
